requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
gunicorn==20.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1
Pillow==8.3.2
torch==1.9.0
torchvision==0.10.0
basicsr==1.3.4.0
realesrgan==0.2.5.0
gfpgan==1.3.4